refactor: split shell_init_test.go into test file and helpers
Extracts shell runner helpers (installShells, stubCampScript*, runBashScript,
runZshScript, runFishScript, ensureCampInPath) into shell_init_helpers_test.go.
Test file drops from 642 to 498 lines, under the 500-line project standard.
